+++ src/GraphicLibs/SDL/Text.hpp
@@ -29,7 +29,7 @@ namespace Arcade {
             public:
                 TextSystem(SDL_Renderer *win,
                            std::vector<std::shared_ptr<Arcade::ECS::IComponent>> &components);
-                void run(float deltaTime,
+                void run(double deltaTime,
                          ECS::IEventManager &eventManager,
                          ECS::IEntityManager &entityManager) final;
             private:
